SVA3D created
the project was created (and hosted on I:/ at SVA).

A ui and server structure was created which allows to pull the current SVASSS, but also sets a structure for future additional dashboards.

As per the current structure, dashboards woudl live WITHIN SVA3D, while SVASSS was kept outside as it is currently.